+++ examples/deploy-example/contracts/Constant.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+contract Constant {
+    uint c = 42;
+    constructor() {}
+
+    function getValue() public view returns (uint) {
+        return c;
+    }
+}