+++ packages/hardhat-zksync-deploy/test/helpers.ts
@@ -1,6 +1,7 @@
 import { resetHardhatContext } from 'hardhat/plugins-testing';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import path from 'path';
+import { ScriptManager } from '../src/script-manager';
 
 declare module 'mocha' {
     interface Context {
@@ -12,8 +13,8 @@ export function useEnvironment(fixtureProjectName: string, networkName = 'hardha
     beforeEach('Loading hardhat environment', function () {
         process.chdir(path.join(__dirname, 'fixture-projects', fixtureProjectName));
         process.env.HARDHAT_NETWORK = networkName;
-
         this.env = require('hardhat');
+        this.scriptManager = new ScriptManager(this.env);
     });
 
     afterEach('Resetting hardhat', function () {